backend/requirements.txt
fastapi
uvicorn
pymongo
qdrant-client
paddlepaddle
paddleocr
PyMuPDF
python-docx
docx2pdf
libreoffice
python-magic
apache-tika
spacy
sentence-transformers
nltk
regex
numpy
scikit-learn
pydantic
python-multipart
docxtpl
